created badge component for messages
