+++ src/interfaces/components/Badge.ts
@@ -1,0 +1,6 @@
+export interface BadgeInterface {
+  content?: React.ReactNode
+  invisible?: boolean
+  variant?: 'standard' | 'dot'
+  children?: React.ReactNode
+}
