+++ src/interfaces/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Button'
 export * from './Input'
 export * from './Paginator'
+export * from './Badge'
